test(kda): hoist KDAAttnBackendForTest shim to test_utils.py
Both KDA test files were carrying an identical copy of the `pool=` →
`recurrent_state_pool=` translation shim added in 466afff7. Move it to
test_utils.py and import from both, dropping the local underscore prefix
since it's now a shared helper.
