+++ python/sgl_jax/test/test_kda_attention.py
@@ -15,34 +15,10 @@ from sgl_jax.srt.managers.schedule_batch import ModelWorkerBatch
 from sgl_jax.srt.mem_cache.recurrent_state_pool import RecurrentStatePool
 from sgl_jax.srt.model_executor.forward_batch_info import ForwardBatch, ForwardMode
 from sgl_jax.srt.utils.mesh_utils import create_device_mesh
+from sgl_jax.test.test_utils import KDAAttnBackendForTest
 
 mesh = create_device_mesh(ici_parallelism=[1, -1], dcn_parallelism=[1, 1])
 jax.sharding.set_mesh(mesh)
-
-
-class _KDAAttnBackendForTest:
-    """Test wrapper that translates `pool=` kwarg to `recurrent_state_pool=`.
-
-    Production routes through HybridLinearAttnBackend, which accepts `pool=`
-    (RadixLinearAttention's call convention) and forwards it to KDA as
-    `recurrent_state_pool=`. These tests assign the raw KDA backend as
-    `forward_batch.attn_backend`, bypassing that wrapper, so we replicate the
-    same translation here.
-    """
-
-    def __init__(self, backend):
-        object.__setattr__(self, "_backend", backend)
-
-    def __call__(self, *args, **kwargs):
-        if "pool" in kwargs:
-            kwargs["recurrent_state_pool"] = kwargs.pop("pool")
-        return self._backend(*args, **kwargs)
-
-    def __getattr__(self, name):
-        return getattr(self._backend, name)
-
-    def __setattr__(self, name, value):
-        setattr(self._backend, name, value)
 
 
 def _scaled_randn(rng: np.random.Generator, shape, scale: float = 0.1) -> np.ndarray:
@@ -348,7 +324,7 @@ def create_test_data(
     extend_prefix_lens = np.zeros(batch_size, dtype=np.int32) if mode == "prefill" else None
     has_initial_state_np = np.asarray(has_initial_state_per_req, dtype=np.bool_)
 
-    backend = _KDAAttnBackendForTest(KDAAttnBackend(mesh=test_mesh))
+    backend = KDAAttnBackendForTest(KDAAttnBackend(mesh=test_mesh))
 
     mwb = ModelWorkerBatch(
         bid=1,
