+++ python/sgl_jax/test/test_kda_attention_dp.py
@@ -16,7 +16,7 @@ from sgl_jax.srt.mem_cache.recurrent_state_pool import RecurrentStatePool
 from sgl_jax.srt.model_executor.forward_batch_info import ForwardBatch, ForwardMode
 from sgl_jax.srt.utils.common_utils import pad_to_bucket
 from sgl_jax.srt.utils.mesh_utils import create_device_mesh
-from sgl_jax.test.test_utils import CustomTestCase
+from sgl_jax.test.test_utils import CustomTestCase, KDAAttnBackendForTest
 
 
 def _scaled_randn(rng: np.random.Generator, shape, scale: float = 0.1) -> np.ndarray:
@@ -25,31 +25,6 @@ def _scaled_randn(rng: np.random.Generator, shape, scale: float = 0.1) -> np.nda
     # Kernel correctness is validated end-to-end (MMLU-pro on tp4dp4, PR #1047);
     # only inputs that grow the state need it (q/k/v/a/b + initial state).
     return rng.standard_normal(shape).astype(np.float32) * scale
-
-
-class _KDAAttnBackendForTest:
-    """Test wrapper that translates `pool=` kwarg to `recurrent_state_pool=`.
-
-    Production routes through HybridLinearAttnBackend, which accepts `pool=`
-    (RadixLinearAttention's call convention) and forwards it to KDA as
-    `recurrent_state_pool=`. These tests assign the raw KDA backend as
-    `forward_batch.attn_backend`, bypassing that wrapper, so we replicate the
-    same translation here.
-    """
-
-    def __init__(self, backend):
-        object.__setattr__(self, "_backend", backend)
-
-    def __call__(self, *args, **kwargs):
-        if "pool" in kwargs:
-            kwargs["recurrent_state_pool"] = kwargs.pop("pool")
-        return self._backend(*args, **kwargs)
-
-    def __getattr__(self, name):
-        return getattr(self._backend, name)
-
-    def __setattr__(self, name, value):
-        setattr(self._backend, name, value)
 
 
 # Reference baselines duplicated from test_kda_attention.py — keep in sync.
@@ -420,7 +395,7 @@ def create_test_data(
     req_pool_indices_cpu = np.arange(total_bs, dtype=np.int32)
 
     real_bs_per_dp = [len(lens_per_rank.get(r, [])) for r in range(dp_size)]
-    backend = _KDAAttnBackendForTest(KDAAttnBackend(mesh=mesh))
+    backend = KDAAttnBackendForTest(KDAAttnBackend(mesh=mesh))
 
     mwb = ModelWorkerBatch(
         bid=1,
